+++ includes/template-tags.php
@@ -259,8 +259,8 @@ function lsx_bd_listing_title( $echo = true ) {
  */
 function lsx_bd_listing_meta( $echo = true ) {
 	$entry_meta = '';
-	$industries = lsx_bd_get_formatted_taxonomy_str( get_the_ID(), 'industry', true );
-	$locations  = lsx_bd_get_formatted_taxonomy_str( get_the_ID(), 'location', true );
+	$industries = get_the_term_list( get_the_ID(), 'industry', '', ', ', '' );
+	$locations  = get_the_term_list( get_the_ID(), 'location', '', ', ', '' );
 	if ( ! empty( $industries ) || ! empty( $locations ) ) {
 		$col_class = '6';
 		if ( empty( $industries ) || empty( $locations ) ) {
@@ -276,18 +276,7 @@ function lsx_bd_listing_meta( $echo = true ) {
 					<span>
 						<i class="fa fa-th"></i>
 						<strong><?php esc_html_e( 'Industry', 'lsx-business-directory' ); ?>: </strong>
-						<?php
-						$count = 0;
-						foreach ( $industries as $industry ) :
-							if ( $count > 0 ) :
-								?>,
-							<?php endif;
-							?>
-								<a href="/industry/<?php echo esc_attr( $industry['slug'] ); ?>"><?php echo esc_attr( $industry['name'] ); ?></a>
-								<?php
-								$count++;
-						endforeach;
-						?>
+						<?php echo wp_kses_post( $industries ); ?>
 					</span>
 				</div>
 				<?php
@@ -301,18 +290,7 @@ function lsx_bd_listing_meta( $echo = true ) {
 					<span>
 						<i class="fa fa-globe"></i>
 						<strong><?php esc_html_e( 'Location', 'lsx-business-directory' ); ?>: </strong>
-						<?php
-						$count = 0;
-						foreach ( $locations as $location ) :
-							if ( $count > 0 ) :
-								?>,
-							<?php endif;
-							?>
-								<a href="/location/<?php echo esc_attr( $location['slug'] ); ?>"><?php echo esc_attr( $location['name'] ); ?></a>
-								<?php
-								$count++;
-						endforeach;
-						?>
+						<?php echo wp_kses_post( $locations ); ?>
 					</span>
 				</div>
 				<?php