+++ changelog.md
@@ -2,6 +2,9 @@
 
 ## 1.1.0 - Unreleased
 
+### Added
+- Added in a "Translations" settings tab to allow the translation of the post type and taxonomy slugs.
+
 ### Fixed
 - Removing the `using-gutenberg` body class when the listing is using blocks.
 - Removing the Bootstrap column classes breaking the padding of the single listings on mobile.